introduce local storage api layer and proper profile store
