+++ app/api/etc/index.js
@@ -5,6 +5,11 @@ import { GenericApiError } from '../common';
 import type { GetSyncProgressResponse } from '../common';
 import type { GetEtcSyncProgressResponse } from './getEtcSyncProgress';
 
+/**
+ * The ETC api layer that handles all requests to the
+ * mantis client which is used as backend for ETC blockchain.
+ */
+
 export const ETC_API_HOST = 'ec2-52-30-28-57.eu-west-1.compute.amazonaws.com';
 export const ETC_API_PORT = 8546;
