+++ app/containers/LoadingPage.js
@@ -29,7 +29,7 @@ export default class LoadingPage extends Component {
       isConnecting, isSyncing, syncPercentage, isLoadingWallets,
       hasBeenConnected, hasBlockSyncingStarted,
     } = stores.ada.networkStatus;
-    const { hasLoadedCurrentLocale, hasLoadedCurrentTheme } = stores.app;
+    const { hasLoadedCurrentLocale, hasLoadedCurrentTheme } = stores.profile;
     return (
       <CenteredLayout>
         <Loading
