+++ features/support/i18n.js
@@ -3,7 +3,7 @@ export default function () {
     this.intl = async (translationId, translationValues = {}) => {
       const translation = await this.client.execute((id, values) => {
         const IntlProvider = require('react-intl').IntlProvider; // eslint-disable-line
-        const locale = daedalus.stores.app.currentLocale;
+        const locale = daedalus.stores.profile.currentLocale;
         const messages = daedalus.translations;
         const intlProvider = new IntlProvider({ locale, messages: messages[locale] }, {});
         return intlProvider.getChildContext().intl.formatMessage({ id }, values);
